refactor(tree): move tree presentation out of mooncake to the CLI seam
CO-6 of the command-output migration: mooncake's tree() no longer writes
directly to stdout. It returns the resolved dependency graph and selected
root, and the moon CLI renders the text tree through CommandOutput.

Behavior-preserving; existing tree snapshots remain unchanged.
