+++ crates/moon/src/cli.rs
@@ -48,6 +48,7 @@ pub(crate) mod search;
 pub(crate) mod shell_completion;
 pub(crate) mod test;
 pub(crate) mod tool;
+pub(crate) mod tree;
 pub(crate) mod update;
 pub(crate) mod upgrade;
 pub(crate) mod version;
@@ -84,6 +85,7 @@ pub(crate) use search::*;
 pub(crate) use shell_completion::*;
 pub(crate) use test::*;
 pub(crate) use tool::*;
+pub(crate) use tree::*;
 pub(crate) use update::*;
 pub(crate) use upgrade::*;
 pub(crate) use version::*;
