+++ crates/moon/src/cli/runtime.rs
@@ -277,7 +277,7 @@ fn dispatch(
         RunWasm(command) => super::run_runwasm(&flags, command, output),
         Search(command) => super::run_search(command, output).map(Into::into),
         Test(command) => super::run_test(flags, command, output).map(Into::into),
-        Tree(command) => super::tree_cli(flags, command, output.user_log()).map(Into::into),
+        Tree(command) => super::tree_cli(flags, command, output).map(Into::into),
         Update(command) => super::update_cli(flags, command, output.user_log()).map(Into::into),
         Upgrade(command) => super::run_upgrade(flags, command).map(Into::into),
         ShellCompletion(command) => super::gen_shellcomp(&flags, command).map(Into::into),
